api/pyproject.toml
[tool.poetry]
name = "django_sensor"
version = "0.0.0"
description = ""
authors = ["Jeff Kerr <jeff@casefleet.com>"]

[tool.poetry.dependencies]
python = "3.8.*"
django = "^3.1.1"
gunicorn = "^20.0.4"
ipython = "^7.18.1"
ipdb = "^0.13.3"
psycopg2-binary = "^2.8.6"
black = "^20.8b1"
django-extensions = "^3.0.8"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-filter = "^2.3.0"
pylint = "^2.6.0"
arrow = "^0.16.0"
influxdb-client = "^1.17.0"
twilio = "^6.59.1"
